% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{sliced.write}
\alias{sliced.write}
\title{Creates external csv/txt files from very large virtual matrices}
\usage{
sliced.write(matrix, file, nblocks, verbose, ...)
}
\arguments{
\item{matrix}{A virtual matrix}

\item{file}{The output file}

\item{nblocks}{Number of blocks}

\item{verbose}{TRUE or FALSE}
}
\description{
Cuts the large virtual matrices into slices and writes them to an external file slice by slice.
             It makes possible that creating very large files on limited system resources.
}
\details{
Function: sliced.write
}
\author{
Samet Tenekeci
}
